+++ src/Core/Managed/Shared/Extensibility/Implementation/TelemetryProcessorChainBuilder.cs
@@ -11,7 +11,7 @@
     public sealed class TelemetryProcessorChainBuilder
     {
         private readonly List<Func<ITelemetryProcessor, ITelemetryProcessor>> factories;
-        private TelemetryConfiguration configuration;
+        private readonly TelemetryConfiguration configuration;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="TelemetryProcessorChainBuilder" /> class.
@@ -19,8 +19,13 @@
         /// <param name="configuration"> The <see cref="TelemetryConfiguration"/> instance to which the constructed processing chain should be set to. </param>        
         public TelemetryProcessorChainBuilder(TelemetryConfiguration configuration)
         {
-            this.factories = new List<Func<ITelemetryProcessor, ITelemetryProcessor>>();
+            if (configuration == null)
+            {
+                throw new ArgumentNullException("configuration");
+            }
+
             this.configuration = configuration;
+            this.factories = new List<Func<ITelemetryProcessor, ITelemetryProcessor>>();
         }
 
         /// <summary>
@@ -42,21 +47,24 @@
         /// </summary>        
         public void Build()
         {
-            var telemetryProcessorChain = new TelemetryProcessorChain();
+            var telemetryProcessorsList = new List<ITelemetryProcessor>();
 
-            // TransmissionProcessor is always appended by default to the end of the chain.
+            // TransmissionProcessor is always appended by default to the end of the chain.            
             ITelemetryProcessor linkedTelemetryProcessor = new TransmissionProcessor(this.configuration);
-            
+            telemetryProcessorsList.Add(linkedTelemetryProcessor);
+
             foreach (var generator in this.factories.AsEnumerable().Reverse())
             {                
                 linkedTelemetryProcessor = generator.Invoke(linkedTelemetryProcessor);
+                telemetryProcessorsList.Add(linkedTelemetryProcessor);
+                
                 if (linkedTelemetryProcessor == null)
                 {
                     throw new InvalidOperationException("TelemetryProcessor returned from TelemetryProcessorFactory cannot be null.");
                 }                
             }
 
-            telemetryProcessorChain.FirstTelemetryProcessor = linkedTelemetryProcessor;
+            var telemetryProcessorChain = new TelemetryProcessorChain(telemetryProcessorsList.AsEnumerable().Reverse());
             this.configuration.TelemetryProcessors = telemetryProcessorChain;
         }
        }