+++ src/Core/Managed/Wp80/Extensibility/Implementation/TelemetryConfigurationFactory.cs
@@ -37,7 +37,7 @@
 
         public virtual void Initialize(TelemetryConfiguration configuration)
         {
-            configuration.ContextInitializers.Add(new SdkVersionPropertyContextInitializer());
+            configuration.TelemetryInitializers.Add(new SdkVersionPropertyTelemetryInitializer());
             configuration.TelemetryInitializers.Add(new TimestampPropertyInitializer());            
 
             // Load customizations from the ApplicationsInsights.config file