+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-# Changelog
+Ôªø# Changelog
 
 This changelog will be used to generate documentation on [release notes page](http://azure.microsoft.com/en-us/documentation/articles/app-insights-release-notes-dotnet/).
 
 ## Version 2.0.0-beta2
-
+- Fix UI thread locking when initializing InMemoryChannel (default channel) from UI thread.
+- Added support for ```ITelemetryProcessor``` and ability to construct chain of TelemetryProcessors via code or config.
+- Version of ```Microsoft.ApplicationInsights.dll``` for the framework 4.6 is now part of the package.
+- IContextInitializer interface is not supported any longer. ContextInitializers collection was removed from TelemetryConfiguraiton object.
 - The max length limit for the ```Name``` property of ```EventTelemetry``` was set to 512.
 - Property ```Name``` of ```OperationContext``` was renamed to ```RootName```
 - Property ```Id``` of ```RequestTelemetry``` was removed.
@@ -42,7 +45,7 @@ This changelog will be used to generate documentation on [release notes page](ht
 
 ## Version 0.15
 
-- New property ```Operation.SyntheticSource``` now available on ```TelemetryContext```. Now you can mark your telemetry items as ìnot a real user trafficî and specify how this traffic was generated. As an example by setting this property you can distinguish traffic from your test automation from load test traffic.
+- New property ```Operation.SyntheticSource``` now available on ```TelemetryContext```. Now you can mark your telemetry items as ‚Äúnot a real user traffic‚Äù and specify how this traffic was generated. As an example by setting this property you can distinguish traffic from your test automation from load test traffic.
 - Channel logic was moved to the separate NuGet called Microsoft.ApplicationInsights.PersistenceChannel. Default channel is now called InMemoryChannel
 - New method ```TelemetryClient.Flush``` allows to flush telemetry items from the buffer synchronously